requirements.txt
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
cattrs==24.1.3
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
Django==5.2
djangorestframework==3.16.0
exceptiongroup==1.2.2
executing==2.2.0
flatbuffers==25.2.10
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.35.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.4
openmeteo_requests==1.4.0
openmeteo_sdk==1.20.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.0
urllib3==2.4.0
wcwidth==0.2.13
